feat(frontend): per-section error boundaries with retry button
- SectionErrorBoundary shows error message + Try Again per section
- Wrap stock, sector, dashboard, trending, rankings routes
- Remove global ErrorBoundary (crash in one section won't kill nav)
