+++ frontend/src/App.jsx
@@ -3,36 +3,9 @@ import routes from "@/routes";
 import React from "react";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { error: null };
-  }
-  static getDerivedStateFromError(error) {
-    return { error };
-  }
-  render() {
-    if (this.state.error) {
-      return (
-        <div style={{ padding: 20, color: "red", fontFamily: "monospace" }}>
-          <h2>Runtime Error</h2>
-          <pre>{this.state.error.message}</pre>
-          <pre>{this.state.error.stack}</pre>
-        </div>
-      );
-    }
-    return this.props.children;
-  }
-}
-
 const App = () => {
   const router = createBrowserRouter(routes);
-
-  return (
-    <ErrorBoundary>
-      <RouterProvider router={router} />
-    </ErrorBoundary>
-  );
+  return <RouterProvider router={router} />;
 };
 
 export default App;
